requirements.txt
websockets
sounddevice
textblob
nltk
numpy
flask
flask-login
flask-cors
aiohttp
requests
watchdog
chromadb
librosa
assemblyai
python-dotenv
gtts
pyttsx3
pymongo
fastapi
uvicorn
pydantic
httpx
google-generativeai
Markupsafe>=2.1.3
Werkzeug>=2.3.7
sentence_transformers
pandas
